+++ et-common-libs/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Provide __assert_func() handler to support libc asserts
 ### Deprecated
 ### Removed
+- Removed old exec_cycles_t structure.
 ### Fixed
 ### Security